backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
ffmpeg-python>=0.2.0
pydantic>=2.5.0
moviepy>=1.0.3
